[DO NOT SUBMIT] Testing multi-use fusion + collapse
The tweak to collapse dims prevents a compilation timeout, but it has
horrible effects on the runtime performance. When there are multiple
reduction ops and it goes down warp reduction, the dispatch has to be
in a very specific state to have good results. Otherwise, compilation
times out or the compiled dispatch is VERY slow (3x total sdxl runtime).

See: https://github.com/iree-org/iree/issues/19868

I found that there are a few sdxl instances of

1 = op with multiple uses
2 = consumer of "1" (transpose)
3 = consumer of "2" (bit extend)

However, there is a reshape that will get stuck between 1-2 or 2-3
depending on which pass you look at (maybe always 2-3). 1-2 could be fused with
multi-use fusion.

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>